+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the "direnv" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Offer to allow blocked envrc when opening it
 ### Changed
 - Improve status item